Rename  _thresholded.py to from thresholded.py
