+++ tests/unit/workflow/test_validators.py
@@ -21,7 +21,6 @@ from typing import Union
 import pytest
 
 from kolena.workflow._datatypes import DataObject
-from kolena.workflow._thresholded import ThresholdedMetrics
 from kolena.workflow._validators import validate_data_object_type
 from kolena.workflow.annotation import BoundingBox
 from kolena.workflow.annotation import Keypoints
@@ -30,6 +29,7 @@ from kolena.workflow.annotation import LabeledPolygon
 from kolena.workflow.annotation import Polygon
 from kolena.workflow.annotation import Polyline
 from kolena.workflow.asset import ImageAsset
+from kolena.workflow.thresholded import ThresholdedMetrics
 
 
 def test__validate_data_object__invalid() -> None:
